KannadaHandwritingRecognition-master/web_app/hwrkannada/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pyhd8ed1ab_0
aom=3.6.1=h63175ca_0
asgiref=3.7.2=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
bottleneck=1.3.5=py37h3a130e4_0
brotli=1.1.0=hcfcfb64_1
brotli-bin=1.1.0=hcfcfb64_1
bzip2=1.0.8=hcfcfb64_5
c-ares=1.27.0=hcfcfb64_0
ca-certificates=2024.2.2=h56e8100_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.2.2=pyhd8ed1ab_0
cloudpickle=2.2.1=pyhd8ed1ab_0
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.12.0=py37hcc03f2d_0
dask-core=2022.2.0=pyhd8ed1ab_0
dav1d=1.2.1=hcfcfb64_0
django=2.0.10=pypi_0
eigen=3.4.0=h91493d7_0
expat=2.5.0=h63175ca_1
ffmpeg=4.2.3=ha925a31_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py37h51bd9d9_0
freeglut=3.2.2=h63175ca_2
freetype=2.12.1=ha860e81_0
fsspec=2023.1.0=pyhd8ed1ab_0
gast=0.5.4=pyhd8ed1ab_0
gettext=0.21.1=h5728263_0
glib=2.78.4=h4dfd0cc_1
glib-tools=2.78.4=h4dfd0cc_1
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.48.1=h568aa7b_1
grpcio=1.48.1=py37h8a48564_1
gst-plugins-base=1.18.5=he07aa86_3
gstreamer=1.18.5=hdff456e_3
h5py=3.3.0=nompi_py37he280515_100
hdf5=1.10.6=nompi_he0bbb20_101
icu=58.2=ha925a31_3
imagecodecs-lite=2019.12.3=py37h0b711f8_5
imageio=2.34.0=pyh4b66e23_0
importlib-metadata=4.11.4=py37h03978a9_0
intel-openmp=2024.0.0=h57928b3_49841
jasper=2.0.33=hc2e4405_1
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_1
keras=2.3.1=py37h21ff451_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.4=py37h8c56517_0
krb5=1.20.1=heb0366b_0
lcms2=2.12=h2a16943_0
lerc=3.0=hd77b12b_0
libabseil=20220623.0=cxx17_h1a56200_6
libaec=1.1.2=h63175ca_1
libblas=3.9.0=21_win64_mkl
libbrotlicommon=1.1.0=hcfcfb64_1
libbrotlidec=1.1.0=hcfcfb64_1
libbrotlienc=1.1.0=hcfcfb64_1
libcblas=3.9.0=21_win64_mkl
libclang=14.0.6=default_h77d9078_1
libclang13=14.0.6=default_h77d9078_1
libcurl=8.1.2=h68f0423_0
libdeflate=1.10=h8ffe710_0
libexpat=2.5.0=h63175ca_1
libffi=3.4.4=hd77b12b_0
libglib=2.78.4=h55e6270_1
libgpuarray=0.7.6=h8ffe710_1003
libhwloc=2.9.1=h51c2c0f_0
libiconv=1.17=hcfcfb64_2
liblapack=3.9.0=21_win64_mkl
liblapacke=3.9.0=21_win64_mkl
libogg=1.3.4=h8ffe710_1
libopencv=4.6.0=haa95532_5
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=3.20.3=h12be248_0
libsqlite=3.45.1=hcfcfb64_0
libssh2=1.11.0=h7dfc565_0
libtiff=4.3.0=hc4061b1_4
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.2=hcfcfb64_1
libwebp-base=1.3.2=h2bbff1b_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.4=hc3477c8_0
libxslt=1.1.37=h0192164_0
libzlib=1.2.13=hcfcfb64_5
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mako=1.3.2=pyhd8ed1ab_0
markdown=3.5.2=pyhd8ed1ab_0
markupsafe=2.1.1=py37hcc03f2d_1
matplotlib=3.5.3=py37h03978a9_2
matplotlib-base=3.5.3=py37hbaab90a_2
mkl=2024.0.0=h66d3029_49657
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=2.6.3=pyhd8ed1ab_1
numexpr=2.7.3=py37h9386db6_2
numpy=1.24.4=pypi_0
olefile=0.47=pyhd8ed1ab_0
opencv=4.6.0=py37hf11a4ad_3
opencv-python=4.9.0.80=pypi_0
openh264=2.3.1=h63175ca_2
openjpeg=2.5.0=hb211442_0
openssl=3.2.1=hcfcfb64_0
packaging=23.2=pyhd8ed1ab_0
pandas=1.3.5=py37h6214cd6_0
partd=1.4.1=pyhd8ed1ab_0
pcre2=10.42=h17e33f8_0
pillow=8.3.2=py37hd7d9ad0_0
pip=24.0=pyhd8ed1ab_0
ply=3.11=py_1
protobuf=3.20.3=py37hd77b12b_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
py-opencv=4.6.0=haa95532_5
pygpu=0.7.6=py37hec80d1f_1003
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.15.4=py37hf2a7229_0
pyqt5-sip=12.9.0=py37hf2a7229_0
pyreadline=2.1=py37h03978a9_1006
pytesseract=0.3.10=pypi_0
python=3.7.0=hc182675_1006
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2024.1=pyhd8ed1ab_0
pywavelets=1.3.0=py37h3a130e4_1
pyyaml=6.0=py37hcc03f2d_4
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
re2=2022.06.01=h0e60522_1
scikit-image=0.19.3=py37h3182a2c_1
scikit-learn=1.0.2=py37hcabfae0_0
scipy=1.7.3=py37hb6553fb_0
setuptools=69.0.3=pyhd8ed1ab_0
sip=6.5.1=py37hf2a7229_2
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=pyhd8ed1ab_0
svt-av1=1.4.1=h63175ca_0
tbb=2021.9.0=h91493d7_0
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h1f41ff6_0
tensorflow-base=1.14.0=py37hc8dfbb8_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=2.3.0=pyhd8ed1ab_0
theano=1.0.5=py37hf2a7229_3
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2020.6.3=py_0
tk=8.6.13=h5226925_1
toml=0.10.2=pyhd8ed1ab_0
toolz=0.12.1=pyhd8ed1ab_0
tornado=6.2=py37hcc03f2d_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=14.0.0=py37hcc03f2d_1
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
vs2017_win-64=19.16.27033=hddac466_18
vswhere=3.1.4=h57928b3_0
werkzeug=2.2.3=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.14.1=py37hcc03f2d_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=hd43e919_0
